format

+++ core/src/infra/entity/block/Block.java
@@ -6,8 +6,8 @@ import infra.entity.Entity;
 
 public abstract class Block extends Entity {
 
-  public static int staticHeight = (int) (Entity.coordinatesScale);
-  public static int staticWidth = (int) (Entity.coordinatesScale);
+  public static int staticHeight = Entity.coordinatesScale;
+  public static int staticWidth = Entity.coordinatesScale;
 
   public Block() {
     super();
@@ -15,7 +15,6 @@ public abstract class Block extends Entity {
     this.zindex = 0;
     this.setHeight(Block.staticHeight);
     this.setWidth(Block.staticWidth);
-    ;
   }
 
   @Override
